exercise_4_multi_model/requirements.txt
# Backend dependencies
fastapi
uvicorn
pydantic
python-dotenv

# Model provider SDKs
groq
huggingface-hub
ollama

# Frontend
streamlit

# HTTP client
httpx
requests